+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## Unreleased: mitmproxy next
 
+* Use newer cryptography APIs to avoid CryptographyDeprecationWarnings.
+  This bumps the minimum required version to cryptography 42.0.
 
 
 ## 06 March 2024: mitmproxy 10.2.3